+++ docs/ChangeLog.md
@@ -211,6 +211,7 @@
 - complete: work around `mawk <= 1.3.4-20230525` type-inference bug (reported by KaKi87) `#D2295` 546499b5
 - main: work around macOS sed (reported by Mossop) `#D2298` a16aa594
 - main: delay attaching in kitty, Ghostty, and VS Code Terminal `#D2215` xxxxxxxx
+- edit: adjust cursor position after `bind -x` in vi_nmap (requested by miltieIV2) `#D2317` xxxxxxxx
 
 ## Contrib